Split up geocode-unit + termops tests into individual test files.
